Refactor: Renames Routes folder to lowercase routes
